requirements.txt
anthropic>=0.39.0
openai>=1.50.0
python-dotenv>=1.0.0
tkinter-tooltip>=3.1.0
